requirements.txt


django-filter==2.3.0
django-rest-swagger==2.2.0


postgis==1.0.4
psycopg2==2.8.5
appdirs==1.4.4
asgiref==3.2.8
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==3.0.7
djangorestframework==3.11.0
drf-yasg==1.17.1
flake8==3.8.3
future==0.18.2
gunicorn==19.9.0
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.4
pathspec==0.8.0
psycopg2-binary==2.8.3
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
ratelim==0.1.6
regex==2020.6.8
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9

